Took functions out of PI and added to utilities folder. Adjusted header links. Added a link to the button on home
