+++ optimizer/src/components/header/Header.js
@@ -3,7 +3,7 @@ import './header.scss';
 import { Link } from 'react-router-dom';
 
 const Header = (props) => {
-    console.log("Header",props);
+    // console.log("Header",props);
     return (
         <header>
             <nav>
