+++ optimizer/src/components/header/header.scss
@@ -11,12 +11,12 @@ header{
             color:#FF5A5F;
         }
         .leftNav{
-            width:60%;
+            width:30%;
             text-align:left;
             margin-left:54px;
         }
         .rightNav{
-            width:40%;
+            width:70%;
             display:flex;
             justify-content: flex-end;
             margin-right:60px;
