+++ optimizer/src/components/home/Home.js
@@ -5,7 +5,7 @@ import { faDollarSign, faHotel } from '@fortawesome/free-solid-svg-icons';
 import {Link} from 'react-router-dom';
 
 const  Home= (props) => {
-    console.log(props);
+    console.log("Home Props",props,localStorage.getItem('token'));
     if(props.match.path==="/logout"){
         localStorage.setItem('token','')
         console.log("NoToken",localStorage.getItem('token'));
@@ -18,7 +18,7 @@ const  Home= (props) => {
                     <h1>Analyze your BnB Listings<br/>with pinpoint geographic analytics</h1>
                 </div>
                 <div className="startButton">
-                    Start Today
+                <Link to="/new-user">Start Today</Link>
                 </div>
             </section>
             <section className="bottomContent">
