+++ optimizer/src/components/home/home.scss
@@ -7,18 +7,20 @@ section{
         box-shadow: 0px 20px 23px -20px rgba(100, 100, 100, 0.25);
         margin-top:100px;
         .startButton{
-            font-family: Raleway;
-            font-style: normal;
-            font-weight: 600;
-            font-size: 15px;
-            line-height: 35px;
-            color: #FFFFFF;
             width: 161px;
             height: 35px;
             background: #FF5A5F;
             box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
             border-radius: 15px;
             margin:0px auto;
+            a{
+                font-family: Raleway;
+                font-style: normal;
+                font-weight: 600;
+                font-size: 15px;
+                line-height: 35px;
+                color: #FFFFFF;
+            }
         }
     }
     &.bottomContent{
